pyproject.toml
[tool.poetry]
name = "fast api"
version = "0.1.0"
description = ""
authors = ["karthikeyan <karthi.sk1467@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.6"
passlib = "^1.7.4"
motor = "^3.6.0"
uvicorn = "^0.34.0"
python-jose = "^3.3.0"
pydantic = {extras = ["email"], version = "^2.10.4"}
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
